+++ game.server/src/card/active/card.bigbbang.effect.ts
@@ -1,18 +1,22 @@
-// cardType = 7
-import { CardType, CharacterStateType } from '../generated/common/enums.js';
-import { removeCard } from '../managers/card.manager.js';
+// cardType = 2
 import {
 	getRoom,
 	getUserFromRoom,
 	saveRoom,
-} from '../utils/room.utils.js';
+	updateCharacterFromRoom,
+} from '../../utils/room.utils';
+import { CardType, CharacterStateType } from '../../generated/common/enums.js';
+import { removeCard } from '../../managers/card.manager.js';
 
-const cardGuerrillaEffect = (roomId: number, userId: string, targetUserId: string): boolean => {
+const cardBigBbangEffect = (roomId: number, userId: string, targetUserId: string): boolean => {
 	const room = getRoom(roomId);
 	const shooter = getUserFromRoom(roomId, userId);
 
-	if (!room || !shooter) return false;
+	if (!room || !shooter) {
+		return false;
+	}
 
+	// "카드 사용을 막아야 하는 상태"만 정의
 	const isBlockedStateUsers = room.users.some(
 		(s) =>
 			s.character &&
@@ -24,12 +28,13 @@ const cardGuerrillaEffect = (roomId: number, userId: string, targetUserId: strin
 		return false;
 	}
 
-	removeCard(shooter, room, CardType.GUERRILLA);
+	removeCard(shooter, room, CardType.BIG_BBANG);
+
 
 	for (let user of room.users) {
 		if (user.character?.stateInfo?.state != null) {
 			if (user.id === userId) {
-				user.character.stateInfo.state = CharacterStateType.GUERRILLA_SHOOTER;
+				user.character.stateInfo.state = CharacterStateType.BIG_BBANG_SHOOTER;
 				user.character.stateInfo.nextState = CharacterStateType.NONE_CHARACTER_STATE;
 				user.character.stateInfo.nextStateAt = `${Date.now() + 10000}`;
 				user.character.stateInfo.stateTargetUserId = targetUserId;
@@ -38,7 +43,7 @@ const cardGuerrillaEffect = (roomId: number, userId: string, targetUserId: strin
 			}
 
 			if (user.character && user.character.hp > 0) {
-				user.character.stateInfo.state = CharacterStateType.GUERRILLA_TARGET;
+				user.character.stateInfo.state = CharacterStateType.BIG_BBANG_TARGET;
 				user.character.stateInfo.nextState = CharacterStateType.NONE_CHARACTER_STATE;
 				user.character.stateInfo.nextStateAt = `${Date.now() + 10000}`;
 				user.character.stateInfo.stateTargetUserId = userId;
@@ -46,8 +51,9 @@ const cardGuerrillaEffect = (roomId: number, userId: string, targetUserId: strin
 		}
 	}
 
-	saveRoom(room);
+	// saveRoom(room);
+
 	return true;
 };
 
-export default cardGuerrillaEffect;
+export default cardBigBbangEffect;
