+++ game.server/src/card/debuff/card.bomb.effect.ts
@@ -1,5 +1,5 @@
 // cardType = 22
-import { getUserFromRoom, updateCharacterFromRoom } from '../utils/room.utils';
+import { getUserFromRoom, updateCharacterFromRoom } from '../../utils/room.utils';
 
 const cardBombEffect = (roomId: number, userId: string, targetUserId: string): boolean => {
 	return true;
