+++ game.server/src/card/equip/card.laser_pointer.effect.ts
@@ -1,7 +1,7 @@
 // cardType = 17
-import { getRoom, getUserFromRoom, updateCharacterFromRoom } from '../utils/room.utils';
-import { CardType } from '../generated/common/enums';
-import { removeCard } from '../managers/card.manager';
+import { getRoom, getUserFromRoom, updateCharacterFromRoom } from '../../utils/room.utils';
+import { CardType } from '../../generated/common/enums';
+import { removeCard } from '../../managers/card.manager';
 
 const cardLaserPointerEffect = (roomId: number, userId: string) : boolean => {
 	const user = getUserFromRoom(roomId, userId);
