+++ game.server/src/card/test/card.bigbbang.effect.test.ts
@@ -3,7 +3,7 @@ import { drawSpecificCard } from '../../managers/card.manager';
 import { Room } from '../../models/room.model';
 import { User } from '../../models/user.model';
 import { getRoom, getUserFromRoom, saveRoom, updateCharacterFromRoom } from '../../utils/room.utils';
-import cardBigBbangEffect from '../card.bigbbang.effect';
+import cardBigBbangEffect from '../active/card.bigbbang.effect';
 
 // Mock dependencies
 jest.mock('../../utils/room.utils');
