+++ game.server/src/card/weapon/card.hand_gun.effect.ts
@@ -1,7 +1,7 @@
 // cardType = 14
-import { getRoom, updateCharacterFromRoom } from '../utils/room.utils';
-import { CardType } from '../generated/common/enums';
-import { removeCard } from '../managers/card.manager.js';
+import { getRoom, updateCharacterFromRoom } from '../../utils/room.utils';
+import { CardType } from '../../generated/common/enums';
+import { removeCard } from '../../managers/card.manager.js';
 
 const cardHandGunEffect = (roomId: number, userId: string): boolean => {
 	try {
