+++ game.server/src/utils/apply.card.effect.ts
@@ -1,28 +1,28 @@
 import { getUserFromRoom, updateCharacterFromRoom } from './room.utils';
 
-import cardAbsorbEffect from '../card/card.absorb.effect';
-import cardAutoRifleEffect from '../card/card.auto_rifle.effect';
-import cardAutoShieldEffect from '../card/card.auto_shield.effect';
-import cardBbangEffect from '../card/card.bbang.effect';
-import cardBigBbangEffect from '../card/card.bigbbang.effect';
-import cardBombEffect from '../card/card.bomb.effect';
-import cardCall119Effect from '../card/card.call_119.effect';
-import cardContainmentUnitEffect from '../card/card.containment_unit.effect';
-import cardDeathMatchEffect from '../card/card.death_match.effect';
-import cardDesertEagleEffect from '../card/card.desert_eagle.effect';
-import cardFleaMarketEffect from '../card/card.flea_market.effect';
-import cardGuerrillaEffect from '../card/card.guerrilla.effect';
-import cardHallucinationEffect from '../card/card.hallucination.effect';
-import cardHandGunEffect from '../card/card.hand_gun.effect';
-import cardLaserPointerEffect from '../card/card.laser_pointer.effect';
-import cardMaturedSavingsEffect from '../card/card.matured_savings.effect';
-import cardRaderEffect from '../card/card.rader.effect';
-import cardSatelliteTargetEffect from '../card/card.satellite_target.effect';
-import cardShieldEffect from '../card/card.shield.effect';
-import cardSniperGunEffect from '../card/card.sniper_gun.effect';
-import cardStealthSuitEffect from '../card/card.stealth_suit.effect';
-import cardVaccineEffect from '../card/card.vaccine.effect';
-import cardWinLotteryEffect from '../card/card.win_lottery.effect';
+import cardAbsorbEffect from '../card/active/card.absorb.effect';
+import cardAutoRifleEffect from '../card/weapon/card.auto_rifle.effect';
+import cardAutoShieldEffect from '../card/equip/card.auto_shield.effect';
+import cardBbangEffect from '../card/active/card.bbang.effect';
+import cardBigBbangEffect from '../card/active/card.bigbbang.effect';
+import cardBombEffect from '../card/debuff/card.bomb.effect';
+import cardCall119Effect from '../card/active/card.call_119.effect';
+import cardContainmentUnitEffect from '../card/debuff/card.containment_unit.effect';
+import cardDeathMatchEffect from '../card/active/card.death_match.effect';
+import cardDesertEagleEffect from '../card/weapon/card.desert_eagle.effect';
+import cardFleaMarketEffect from '../card/active/card.flea_market.effect';
+import cardGuerrillaEffect from '../card/active/card.guerrilla.effect';
+import cardHallucinationEffect from '../card/active/card.hallucination.effect';
+import cardHandGunEffect from '../card/weapon/card.hand_gun.effect';
+import cardLaserPointerEffect from '../card/equip/card.laser_pointer.effect';
+import cardMaturedSavingsEffect from '../card/active/card.matured_savings.effect';
+import cardRaderEffect from '../card/equip/card.rader.effect';
+import cardSatelliteTargetEffect from '../card/debuff/card.satellite_target.effect';
+import cardShieldEffect from '../card/active/card.shield.effect';
+import cardSniperGunEffect from '../card/weapon/card.sniper_gun.effect';
+import cardStealthSuitEffect from '../card/equip/card.stealth_suit.effect';
+import cardVaccineEffect from '../card/active/card.vaccine.effect';
+import cardWinLotteryEffect from '../card/active/card.win_lottery.effect';
 import { repeatDeck } from '../managers/card.manager';
 
 // 카드 효과 적용 함수
