+++ game.server/src/utils/test/apply.card.effect.test.ts
@@ -3,29 +3,29 @@ import { getUserFromRoom, updateCharacterFromRoom } from '../room.utils';
 import { repeatDeck } from '../../managers/card.manager.js';
 
 // 카드 효과 모듈 mock
-import cardAbsorbEffect from '../../card/card.absorb.effect.js';
-import cardAutoRifleEffect from '../../card/card.auto_rifle.effect.js';
-import cardAutoShieldEffect from '../../card/card.auto_shield.effect.js';
-import cardBbangEffect from '../../card/card.bbang.effect.js';
-import cardBigBbangEffect from '../../card/card.bigbbang.effect.js';
-import cardBombEffect from '../../card/card.bomb.effect.js';
-import cardCall119Effect from '../../card/card.call_119.effect.js';
-import cardContainmentUnitEffect from '../../card/card.containment_unit.effect.js';
-import cardDeathMatchEffect from '../../card/card.death_match.effect.js';
-import cardDesertEagleEffect from '../../card/card.desert_eagle.effect.js';
-import cardFleaMarketEffect from '../../card/card.flea_market.effect.js';
-import cardGuerrillaEffect from '../../card/card.guerrilla.effect.js';
-import cardHallucinationEffect from '../../card/card.hallucination.effect.js';
-import cardHandGunEffect from '../../card/card.hand_gun.effect.js';
-import cardLaserPointerEffect from '../../card/card.laser_pointer.effect.js';
-import cardMaturedSavingsEffect from '../../card/card.matured_savings.effect.js';
-import cardRaderEffect from '../../card/card.rader.effect.js';
-import cardSatelliteTargetEffect from '../../card/card.satellite_target.effect.js';
-import cardShieldEffect from '../../card/card.shield.effect.js';
-import cardSniperGunEffect from '../../card/card.sniper_gun.effect.js';
-import cardStealthSuitEffect from '../../card/card.stealth_suit.effect.js';
-import cardVaccineEffect from '../../card/card.vaccine.effect.js';
-import cardWinLotteryEffect from '../../card/card.win_lottery.effect.js';
+import cardAbsorbEffect from '../../card/active/card.absorb.effect.js';
+import cardAutoRifleEffect from '../../card/weapon/card.auto_rifle.effect.js';
+import cardAutoShieldEffect from '../../card/equip/card.auto_shield.effect.js';
+import cardBbangEffect from '../../card/active/card.bbang.effect.js';
+import cardBigBbangEffect from '../../card/active/card.bigbbang.effect.js';
+import cardBombEffect from '../../card/debuff/card.bomb.effect.js';
+import cardCall119Effect from '../../card/active/card.call_119.effect.js';
+import cardContainmentUnitEffect from '../../card/debuff/card.containment_unit.effect.js';
+import cardDeathMatchEffect from '../../card/active/card.death_match.effect.js';
+import cardDesertEagleEffect from '../../card/weapon/card.desert_eagle.effect.js';
+import cardFleaMarketEffect from '../../card/active/card.flea_market.effect.js';
+import cardGuerrillaEffect from '../../card/active/card.guerrilla.effect.js';
+import cardHallucinationEffect from '../../card/active/card.hallucination.effect.js';
+import cardHandGunEffect from '../../card/weapon/card.hand_gun.effect.js';
+import cardLaserPointerEffect from '../../card/equip/card.laser_pointer.effect.js';
+import cardMaturedSavingsEffect from '../../card/active/card.matured_savings.effect.js';
+import cardRaderEffect from '../../card/equip/card.rader.effect.js';
+import cardSatelliteTargetEffect from '../../card/debuff/card.satellite_target.effect.js';
+import cardShieldEffect from '../../card/active/card.shield.effect.js';
+import cardSniperGunEffect from '../../card/weapon/card.sniper_gun.effect.js';
+import cardStealthSuitEffect from '../../card/equip/card.stealth_suit.effect.js';
+import cardVaccineEffect from '../../card/active/card.vaccine.effect.js';
+import cardWinLotteryEffect from '../../card/active/card.win_lottery.effect.js';
 
 jest.mock('../room.utils', () => ({
 	getUserFromRoom: jest.fn(),
